Fix organization cases for both organization admin and members
1. This commit makes fixes so that the organization cases count and search view works for both organization administrator and its members.
2. Previously, it was only working for the organization administrator.
3. However, this inlcudes new complexity because of the need to figure out:
a. is the user an organization admin
b. is the user a member of an organization
if a is true, then pick out other users in the organization
if b is true, then find the organization and do a reverse lookup to find members in the organization.
Note: this possibly needs a table relationship design rethink because of this unnecessary complexity. Will need to think of it soon.